Refactor KFP pipeline implementation
