+++ elyra/pipeline/component.py
@@ -228,7 +228,7 @@ class ComponentParser(LoggingConfigurable):  # ABC
     _file_types: List[str] = None
     _parser_class_map: Dict[str, str] = {
         "APACHE_AIRFLOW": "elyra.pipeline.airflow.airflow_component_parser:AirflowComponentParser",
-        "KUBEFLOW_PIPELINES": "elyra.pipeline.kfp.component_parser_kfp:KfpComponentParser",
+        "KUBEFLOW_PIPELINES": "elyra.pipeline.kfp.kfp_component_parser:KfpComponentParser",
     }
 
     @classmethod
