+++ elyra/tests/pipeline/test_processor.py
@@ -18,7 +18,7 @@ import tarfile
 
 import pytest
 
-from elyra.pipeline.kfp.processor_kfp import KfpPipelineProcessor
+from elyra.pipeline.kfp.kfp_processor import KfpPipelineProcessor
 from elyra.pipeline.pipeline import GenericOperation
 from elyra.pipeline.properties import ElyraProperty
 from elyra.pipeline.registry import PipelineProcessorRegistry
